% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraday.R
\name{main}
\alias{main}
\title{Intraday Main Function}
\usage{
main(date, animal, venueName)
}
\arguments{
\item{date}{Date in unix format YYYY-MM-DD}

\item{animal}{Venue type e.g. THOROUGHBRED, HARNESS, GREYHOUND}

\item{venueName}{name of venue (uppercase)}
}
\description{
This function allows you to express your love of cats.
}
\examples{
main('2016-10-26','THOROUGHBRED','HAPPY VALLEY')
}
\keyword{intraday}

